data loader moved, bugs in linear encoder and san fixed, scripts added

+++ src/encoder/linEnc.py
@@ -19,7 +19,7 @@ class LinearEdgeEncoder(torch.nn.Module):
         self.lin = torch.nn.Linear(input_channels, output_channels)
         
     def forward(self, batch):
-        batch.edge_attr = self.lin(batch.edge_attr)
+        batch.edge_attr = self.lin(batch.edge_attr.type(torch.float32))
         return batch
